Separate utility functions into separate JS
